Break apart the code more
