+++ src/utils/index.ts
@@ -1,3 +1,4 @@
-export * from './entities'
 export * from './blocklist'
+export * from './entities'
+export * from './results'
 export * from './tree'
