+++ src/utils/tree.ts
@@ -1,4 +1,9 @@
-import type { AnalyzedEntity, AnyAnalyzedFile } from '../types'
+import type {
+  AnalyzedEntity,
+  AnyAnalyzedFile,
+  FullyAnalyzedEntity,
+  FullyAnalyzedFile,
+} from '../types'
 
 export const flattenEntireTree = <T extends AnyAnalyzedFile>(
   items: AnalyzedEntity[]
@@ -13,4 +18,25 @@ export const flattenEntireTree = <T extends AnyAnalyzedFile>(
   })
 
   return flattened
+}
+
+export const getResultsAsList = (
+  analyzedEntities: FullyAnalyzedEntity[],
+  limit?: number
+): FullyAnalyzedFile[] => {
+  const flatFileResults: FullyAnalyzedFile[] = flattenEntireTree<
+    FullyAnalyzedFile
+  >(analyzedEntities)
+    .filter((entity) => {
+      return entity.type === 'file' && !!entity.complexityReport
+    })
+    // Sort by codehawk score, ascending (most complex files are first in the list)
+    .sort((entityA, entityB) => {
+      return (
+        entityA.complexityReport.codehawkScore -
+        entityB.complexityReport.codehawkScore
+      )
+    })
+
+  return limit ? flatFileResults.slice(0, limit) : flatFileResults
 }
